+++ sdk/core/Azure.Core/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## 1.4.0-preview.1 (Unreleased)
 
+### Added
+- Added `ObjectSerializer` base class for serialization.
+- Added `IMemberNameConverter` for converting member names to serialized property names.
+- Added `JsonObjectSerializer` that implements `ObjectSerializer` for `System.Text.Json`.
+
 ### Fixed
 - Connection leak for retried non-buffered requests on .NET Framework.